Moves some methods into a new Monolog module.
The method in src/reader.rs that read a file content to get a vector of
log::Log is now part of the input::monolog::Monolog struct.
All the structs that are used to import logs and get a vector of
log::Log have to implement the trait input::Importable.